requirements.txt
loguru==0.7.3
python-dotenv==1.0.1
Requests==2.32.3
